# api/openapi.yaml
openapi: 3.0.3
info:
  title: Blog API
  description: |
    RESTful API for multi-tenant blog platform.

    ## Authentication

    All endpoints (except `/api/health`) require API key authentication.

    Include your API key in one of these headers:
    - `Authorization: Bearer <your-api-key>`
    - `X-API-Key: <your-api-key>`

    ## API Key Types

    - **Admin keys** (`sa_`): Full access to all resources
    - **User keys** (`sk_`): Access to owned/member sites
    - **Site keys** (`ss_`): Access limited to specific site

    ## Scopes

    - `read`: Read-only access
    - `write`: Create and update resources
    - `delete`: Delete resources
    - `admin`: Full administrative access
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in X-API-Key header
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error details
      required:
        - error
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        site_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TagCreate:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/sites/{siteId}/tags:
    parameters:
      - name: siteId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Tags
      summary: List tags
      description: List all tags for a site
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
                  count:
                    type: integer

    post:
      tags:
        - Tags
      summary: Create tag
      description: Create a new tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
        '409':
          description: Conflict - slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tags/{tagId}:
    parameters:
      - name: tagId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags:
        - Tags
      summary: Update tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'

    delete:
      tags:
        - Tags
      summary: Delete tag
      responses:
        '204':
          description: Tag deleted
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Users
...